requirements.txt
Flask==3.0.2
openai==1.16.1
PyYAML==6.0.1
requests==2.31.0
lxml==5.2.1